+++ app.js
@@ -13,6 +13,7 @@ var thumb = require("./backend/thumb");
 var settings = require("./backend/settings");
 var directory = require("./backend/directory");
 var streamer = require("./backend/streamer");
+var monitor = require("./backend/monitor");
 
 var app = express();
 
@@ -34,6 +35,8 @@ app.use("/backend/thumb", thumb);
 app.use("/backend/settings", settings);
 app.use("/backend/directory", directory);
 app.use("/backend/streamer", streamer);
+app.use("/backend/monitor", monitor);
+
 app.use("/webhook", webhookRouter);
 app.use("/*", uiRouter);